feat(client-app): address bar server component
